fastboot: Drop support for downloading to buffer
The option to download to a buffer instead of a file was introduced
because in some workloads it is required to have a contiguous image
in memory. With recent changes now ramfs can provide such a buffer
via memmap API even when it downloaded the data to a file. This makes
the explicit download to buffer option unnecessary, so remove it.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>