tlsf: Add tracking of added tlsf memory pools
When configured to use the TLSF allocator, the malloc_stats function
only walks the initial memory pool, as that is the only pool it is aware
of, and TLSF doesn't link pools together in a way that the walker can
follow.

Add a wrapper function around tlsf_add_pool to add additional tracking
of added pools, so that they can be walked by the malloc_stats function
and meminfo can report accurate heap utilization.

Signed-off-by: David Dgien <dgienda125@gmail.com>
Reviewed-by: Tyler Reece <mtreece@acm.org>